+++ src/index.ts
@@ -7,6 +7,9 @@ import { stopCommand } from "./cli/stop";
 import { proxyCommand } from "./cli/proxy";
 import { doctorCommand } from "./cli/doctor";
 import { caCommand } from "./cli/root";
+import { daemonCommand } from "./cli/daemon";
+import { forwardCommand } from "./cli/forward";
+import { shareCommand, tunnelServerCommand } from "./cli/tunnel";
 
 const program = new Command();
 
@@ -21,5 +24,9 @@ program.addCommand(stopCommand);
 program.addCommand(proxyCommand);
 program.addCommand(doctorCommand);
 program.addCommand(caCommand);
+program.addCommand(daemonCommand);
+program.addCommand(forwardCommand);
+program.addCommand(shareCommand);
+program.addCommand(tunnelServerCommand);
 
 program.parse();